+++ src/credits/selectors.js
@@ -1,8 +1,8 @@
 import { SelectorFactory } from 'ethical-jobs-redux';
 import { createSelector } from 'reselect';
 
-export const fetchingSelector = SelectorFactory.create('credits', 'fetching');
+export const fetching = SelectorFactory.create('credits', 'fetching');
 
-export const errorSelector = SelectorFactory.create('credits', 'error');
+export const error = SelectorFactory.create('credits', 'error');
 
-export const creditPacksSelector = state => state.getIn(['entities', 'credits', 'creditPacks']);
+export const creditPacks = state => state.getIn(['entities', 'credits', 'creditPacks']);