+++ src/invoices/selectors.js
@@ -2,23 +2,23 @@ import { SelectorFactory } from 'ethical-jobs-redux';
 import { createSelector } from 'reselect';
 import selectByFilters from './filters';
 
-export const fetchingSelector = SelectorFactory.create('invoices', 'fetching');
+export const fetching = SelectorFactory.create('invoices', 'fetching');
 
-export const errorSelector = SelectorFactory.create('invoices', 'error');
+export const error = SelectorFactory.create('invoices', 'error');
 
-export const filtersSelector = SelectorFactory.createFiltersSelector('invoices');
+export const filters = SelectorFactory.createFiltersSelector('invoices');
 
-export const resultSelector = SelectorFactory.createResultSelector('invoices');
+export const result = SelectorFactory.createResultSelector('invoices');
 
-export const resultsSelector = SelectorFactory.createResultsSelector('invoices');
+export const results = SelectorFactory.createResultsSelector('invoices');
 
-export const invoicesSelector = SelectorFactory.createEntitiesSelector('invoices');
+export const invoices = SelectorFactory.createEntitiesSelector('invoices');
 
-export const orderedInvoicesSelector = SelectorFactory.createOrderedEntitiesSelector(invoicesSelector, resultsSelector);
+export const orderedInvoices = SelectorFactory.createOrderedEntitiesSelector(invoices, results);
 
-export const invoiceByIdSelector = SelectorFactory.createIdSelector(invoicesSelector, resultSelector);
+export const invoiceByResult = SelectorFactory.createIdSelector(invoices, result);
 
 export const invoicesByFiltersSelector = createSelector(
-  [orderedInvoicesSelector, filtersSelector],
+  [orderedInvoices, filters],
   selectByFilters
 );